+++ mailpoet/lib/Captcha/Validator/RecaptchaValidator.php
@@ -1,10 +1,10 @@
 <?php // phpcs:ignore SlevomatCodingStandard.TypeHints.DeclareStrictTypes.DeclareStrictTypesMissing
 
-namespace MailPoet\Subscription\Captcha\Validator;
+namespace MailPoet\Captcha\Validator;
 
 use MailPoet\Captcha\ReCaptchaValidator as Validator;
 
-class RecaptchaValidator implements CaptchaValidator {
+class RecaptchaValidator {
 
   /** @var Validator */
   private $validator;
