+++ mailpoet/lib/Subscribers/SubscriberSubscribeController.php
@@ -2,6 +2,11 @@
 
 namespace MailPoet\Subscribers;
 
+use MailPoet\Captcha\CaptchaConstants;
+use MailPoet\Captcha\CaptchaSession;
+use MailPoet\Captcha\Validator\CaptchaValidator;
+use MailPoet\Captcha\Validator\RecaptchaValidator;
+use MailPoet\Captcha\Validator\ValidationError;
 use MailPoet\Entities\FormEntity;
 use MailPoet\Entities\SubscriberEntity;
 use MailPoet\Entities\SubscriberTagEntity;
@@ -11,11 +16,6 @@ use MailPoet\NotFoundException;
 use MailPoet\Segments\SubscribersFinder;
 use MailPoet\Settings\SettingsController;
 use MailPoet\Statistics\StatisticsFormsRepository;
-use MailPoet\Subscription\Captcha\CaptchaConstants;
-use MailPoet\Subscription\Captcha\CaptchaSession;
-use MailPoet\Subscription\Captcha\Validator\CaptchaValidator;
-use MailPoet\Subscription\Captcha\Validator\RecaptchaValidator;
-use MailPoet\Subscription\Captcha\Validator\ValidationError;
 use MailPoet\Subscription\Throttling as SubscriptionThrottling;
 use MailPoet\Tags\TagRepository;
 use MailPoet\UnexpectedValueException;
