+++ mailpoet/tests/acceptance/Settings/ConfirmConfirmationPageCest.php
@@ -3,15 +3,15 @@
  namespace MailPoet\Test\Acceptance;
 
  use Codeception\Util\Locator;
- use MailPoet\Subscription\Captcha\CaptchaConstants;
+ use MailPoet\Captcha\CaptchaConstants;
  use MailPoet\Test\DataFactories\Form;
  use MailPoet\Test\DataFactories\Settings;
 
-/**
+ /**
  * @group frontend
  */
 class ConfirmConfirmationPageCest {
-  
+
     const CONFIRMATION_MESSAGE_TIMEOUT = 20;
     const FORM_NAME = 'Subscription Acceptance Test Form';
