+++ mailpoet/tests/integration/API/JSON/v1/SubscribersTest.php
@@ -9,6 +9,8 @@ use MailPoet\API\JSON\Response as APIResponse;
 use MailPoet\API\JSON\ResponseBuilders\SubscribersResponseBuilder;
 use MailPoet\API\JSON\SuccessResponse;
 use MailPoet\API\JSON\v1\Subscribers;
+use MailPoet\Captcha\CaptchaConstants;
+use MailPoet\Captcha\CaptchaSession;
 use MailPoet\DI\ContainerWrapper;
 use MailPoet\Entities\CustomFieldEntity;
 use MailPoet\Entities\FormEntity;
@@ -28,8 +30,6 @@ use MailPoet\Subscribers\SubscriberSaveController;
 use MailPoet\Subscribers\SubscribersRepository;
 use MailPoet\Subscribers\SubscriberSubscribeController;
 use MailPoet\Subscribers\SubscriberTagRepository;
-use MailPoet\Subscription\Captcha\CaptchaConstants;
-use MailPoet\Subscription\Captcha\CaptchaSession;
 use MailPoet\Tags\TagRepository;
 use MailPoet\Test\DataFactories\CustomField as CustomFieldFactory;
 use MailPoet\Test\DataFactories\DynamicSegment;
